% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{batch_simulation}
\alias{batch_simulation}
\title{Run a batch of simulations in a console interface}
\usage{
batch_simulation(
  total_time,
  map_filename = system.file("extdata", "map.csv", package = "li19trustmodel"),
  config = system.file("extdata", "params.json", package = "li19trustmodel"),
  num_adversaries = c(0, 2, 5, 8, 10),
  adversary_types = c("BadMouther", "ContextSetter"),
  colours = c("blue", "red", "green", "orange", "purple")
)
}
\description{
Simulate the trust model and the mobile network and iterate for an amount
of time for differing amounts of adversaries
}
\keyword{model}
\keyword{run}
\keyword{simulate}
\keyword{simulation}
\keyword{trust}
