% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{run_gui}
\alias{run_gui}
\title{Run the simulation in a graphical interface}
\usage{
run_gui(
  map_filename = system.file("extdata", "map.csv", package = "li19trustmodel"),
  config = system.file("extdata", "params.json", package = "li19trustmodel")
)
}
\description{
Simulate the trust model and the mobile network and iterate
}
\keyword{model}
\keyword{run}
\keyword{simulate}
\keyword{simulation}
\keyword{trust}
