% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Run the simulation in a console interface}
\usage{
run_simulation(
  total_time,
  map_filename = system.file("extdata", "map.csv", package = "li19trustmodel"),
  config = system.file("extdata", "params.json", package = "li19trustmodel"),
  write_plots = TRUE
)
}
\description{
Simulate the trust model and the mobile network and iterate for an amount
of time
}
\keyword{model}
\keyword{run}
\keyword{simulate}
\keyword{simulation}
\keyword{trust}
